ui_edited

+++ kiem-tra-module-4/src/main/java/com/codegym/kiemtramodule4/repository/NationRepository.java
@@ -3,5 +3,5 @@ package com.codegym.kiemtramodule4.repository;
 import com.codegym.kiemtramodule4.model.Nation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NationRepository extends JpaRepository<Nation, Long> {
+public interface NationRepository extends JpaRepository<Nation, Integer> {
 }
